pyproject.toml
[project]
name = "dm-keith"
version = "0.1.0"
description = "Dungeon Master Keith — sarcastic dungeon master AI persona with Telegram bot integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.40.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "tenacity>=8.2",
    "typing-extensions>=4.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "black>=24.4",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
